base backend configuration
feat: add base configuration, routes and mongoose setup to the index.js
