+++ server/index.js
@@ -1,0 +1,48 @@
+import express from 'express'
+import bodyParser from 'body-parser'
+import mongoose from 'mongoose'
+import cors from 'cors'
+import dotenv from 'dotenv'
+import helmet from 'helmet'
+import morgan from 'morgan'
+import clientRoutes from './routes/client.js'
+import generalRoutes from './routes/general.js'
+import managementRoutes from './routes/management.js'
+import salesRoutes from './routes/sales.js'
+
+/* CONFIGURATION */
+dotenv.config()
+const app = express()
+app.use(express.json())
+app.use(helmet())
+app.use(helmet.crossOriginResourcePolicy({ policy: 'cross-origin' }))
+app.use(morgan('common'))
+app.use(bodyParser.json())
+app.use(bodyParser.urlencoded({ extended: false }))
+app.use(cors())
+
+/* ROUTES */
+app.use('/client', clientRoutes)
+app.use('/general', generalRoutes)
+app.use('/management', managementRoutes)
+app.use('/sales', salesRoutes)
+
+/*MONGOOSE SETUP*/
+const PORT = process.env.PORT || 9000;
+mongoose
+  .connect(process.env.MONGO_URL, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+  })
+  .then(() => {
+    app.listen(PORT, () => console.log(`Server Port: ${PORT}`));
+
+    /* ONLY ADD DATA ONE TIME */
+    //  AffiliateStat.insertMany(dataAffiliateStat);
+    //  OverallStat.insertMany(dataOverallStat);
+    //  Product.insertMany(dataProduct);
+    //  ProductStat.insertMany(dataProductStat);
+    //  Transaction.insertMany(dataTransaction);
+    //  User.insertMany(dataUser);
+  })
+  .catch((error) => console.log(`${error} did not connect`));
